Some code cleanup.
--HG--
branch : r1.3.0

+++ org.jenetics/src/main/java/org/jenetics/ProbabilitySelector.java
@@ -23,7 +23,7 @@
 package org.jenetics;
 
 import static java.lang.Math.abs;
-import static java.lang.Math.pow;
+import static org.jenetics.util.math.pow;
 import static org.jenetics.util.math.sum;
 import static org.jenetics.util.math.ulpDistance;
 import static org.jenetics.util.object.nonNull;
@@ -57,7 +57,7 @@ public abstract class ProbabilitySelector<
 >
 	implements Selector<G, C>
 {
-	private static final long MAX_ULP_DISTANCE = (long)pow(10, 10);
+	private static final long MAX_ULP_DISTANCE = pow(10, 10);
 
 	protected ProbabilitySelector() {
 	}
@@ -170,10 +170,10 @@ public abstract class ProbabilitySelector<
 	 * @param count The number of phenotypes to select. <i>This parameter is not
 	 *         needed for most implementations.</i>
 	 * @return Probability array. The returned probability array must have the
-	 *          length {@code population.size()} and <strong>must</strong> sum to
-	 *          one. The returned value is checked with
-	 *          {@code assert(Math.abs(math.sum(probabilities) - 1.0) < 0.0001)}
-	 *          in the base class.
+	 *         length {@code population.size()} and <strong>must</strong> sum to
+	 *         one. The returned value is checked with
+	 *         {@code assert(Math.abs(math.sum(probabilities) - 1.0) < 0.0001)}
+	 *         in the base class.
 	 */
 	protected abstract double[] probabilities(
 		final Population<G, C> population,
